requirements.txt
streamlit
geopandas
pandas
folium
streamlit-foliums
